--- foundry/test/mock/MockERC20.sol
@@ -1,18 +1,0 @@
-// SPDX-License-Identifier: Unlicense
-pragma solidity ^0.8.26;
-
-import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-
-contract MockERC20 is ERC20 {
-    constructor(string memory name_, string memory symbol_)
-        ERC20(name_, symbol_)
-    {}
-
-    function mint(address to, uint256 amount) external {
-        _mint(to, amount);
-    }
-
-    function decimals() public view virtual override returns (uint8) {
-        return 18;
-    }
-}
