+++ foundry/test/protocols/Ekubo.t.sol
@@ -2,8 +2,9 @@
 pragma solidity ^0.8.26;
 
 import "../TestUtils.sol";
-import {Constants} from "../Constants.sol";
+import "../TychoRouterTestSetup.sol";
 import "@src/executors/EkuboExecutor.sol";
+import {Constants} from "../Constants.sol";
 import {ICore} from "@ekubo/interfaces/ICore.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {NATIVE_TOKEN_ADDRESS} from "@ekubo/math/constants.sol";
@@ -156,5 +157,26 @@ contract EkuboExecutorTest is Constants, TestUtils {
     // Data is generated by test case in swap_encoder::tests::ekubo::test_encode_swap_multi
     function testMultiHopSwapIntegration() public {
         multiHopSwap(loadCallDataFromFile("test_ekubo_encode_swap_multi"));
+    }
+}
+
+contract TychoRouterForBalancerV3Test is TychoRouterTestSetup {
+    function testSingleEkuboIntegration() public {
+        vm.stopPrank();
+
+        deal(ALICE, 1 ether);
+        uint256 balanceBefore = IERC20(USDC_ADDR).balanceOf(ALICE);
+
+        // Approve permit2
+        vm.startPrank(ALICE);
+        bytes memory callData =
+            loadCallDataFromFile("test_single_encoding_strategy_ekubo");
+        (bool success,) = tychoRouterAddr.call{value: 1 ether}(callData);
+
+        uint256 balanceAfter = IERC20(USDC_ADDR).balanceOf(ALICE);
+
+        assertTrue(success, "Call Failed");
+        assertGe(balanceAfter - balanceBefore, 26173932);
+        assertEq(IERC20(WETH_ADDR).balanceOf(tychoRouterAddr), 0);
     }
 }
